Add 'Shared Element Transitions' topic to documentation
A new topic called 'Shared Element Transitions' has been included in the Compose module in our documentation. This topic comprehensively covers the use of Shared Element Transitions in Jetpack Compose, highlighting the problems with existing implementations and proposing a solution that enhances maintainability. The 'Shared Element Transitions' has also been added to the table of contents.